visitdata/requirements.txt
apache-airflow==1.10.6
boto3==1.10.33
botocore==1.13.33
psycopg2==2.8.4
python-dotenv==0.10.3
pylint==2.4.4
SQLAlchemy==1.3.11
sqlalchemy-redshift==0.7.5
sphinx-autoapi==1.2.1